% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_datoformat.R
\name{textfunc.datoformat}
\alias{textfunc.datoformat}
\title{Converts a Date into date format}
\usage{
textfunc.datoformat(date)
}
\arguments{
\item{date}{Date("date"), POSIX("date"), etc.; Date to convert}
}
\value{
Character string in date format "%e. %B %Y"
}
\description{
Converts a Date into date format
}
