% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_getSentimentScore.R
\name{textfunc.getSentimentFromScore}
\alias{textfunc.getSentimentFromScore}
\title{Get sentiment from sentimen score}
\usage{
textfunc.getSentimentFromScore(SentimentScore)
}
\arguments{
\item{SentimentScore}{integer; The Sentiment Score you want sentiment text on}
}
\value{
Sentiment as character - positive, neutral og negative
}
\description{
Get sentiment from sentimen score
}
