% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_getSentimentScore.R
\name{textfunc.getSentimentScore}
\alias{textfunc.getSentimentScore}
\title{Get sentiment from text}
\usage{
textfunc.getSentimentScore(textdata, cleanText = TRUE)
}
\arguments{
\item{textdata}{character vector; The text you want to get sentiment on}

\item{cleanText}{logical; If TRUE it will clean text, before getting sentiment}
}
\value{
Sentiment score as integer
}
\description{
Get sentiment from text
}
