% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_oplist.R
\name{textfunc.listtext}
\alias{textfunc.listtext}
\title{Comma separat a list of text with "and" in the end}
\usage{
textfunc.listtext(text, lastsep = "og")
}
\arguments{
\item{text}{c("text"); List of character}

\item{lastsep}{character("og"); The last separator to use}
}
\value{
Character string
}
\description{
Comma separat a list of text with "and" in the end
}

