% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_numeral.R
\name{textfunc.numeral}
\alias{textfunc.numeral}
\title{Trun a number into a numeral (Danish)}
\usage{
textfunc.numeral(number, ordinal = FALSE, et = FALSE)
}
\arguments{
\item{number}{numeric(number) or integer(number); Number to convert to a numeral}

\item{ordinal}{logical(); If the number should be converted into an ordinal}

\item{et}{logical(); If "et" should be used insted of "en"}
}
\value{
Character
}
\description{
Trun a number into a numeral (Danish)
}

