% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfunc_textcleaner.R
\name{textfunc.textcleaner}
\alias{textfunc.textcleaner}
\title{Clean SoMe text}
\usage{
textfunc.textcleaner(textdata, removeRT = FALSE, removeHashtag = FALSE,
  removeHandles = FALSE)
}
\arguments{
\item{textdata}{character vector; The text you want cleaned}

\item{removeRT}{logical; If TRUE it wil remove 'RT' and 'via' from text}

\item{removeHashtag}{logical; If TRUE it wil remove all words begining with '#' from text}

\item{removeHandles}{logical; If TRUE it wil remove all words begining with '@' from text}
}
\value{
Character string
}
\description{
Clean SoMe text
}
